Crate ComparableByName interface to standardise 'looser comparison' across entity classes
