+++ src/main/java/seedu/address/model/ComparableByName.java
@@ -1,0 +1,12 @@
+package seedu.address.model;
+
+public interface ComparableByName<T> {
+
+    /**
+     * This defines a lower level of equality - comparison by name/description
+     * for relevant objects that can be compared by such.
+     * @param other other object
+     * @return true if the objects have the same name.
+     */
+    boolean hasSameName(T other);
+}
