+++ src/main/java/seedu/address/model/client/UniquePersonList.java
@@ -1,137 +1,137 @@
-package seedu.address.model.client;
-
-import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
-
-import java.util.Iterator;
-import java.util.List;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import seedu.address.model.client.exceptions.DuplicatePersonException;
-import seedu.address.model.client.exceptions.PersonNotFoundException;
-
-/**
- * A list of persons that enforces uniqueness between its elements and does not allow nulls.
- * A client is considered unique by comparing using {@code Person#isSamePerson(Person)}. As such, adding and updating of
- * persons uses Person#isSamePerson(Person) for equality so as to ensure that the client being added or updated is
- * unique in terms of identity in the UniquePersonList. However, the removal of a client uses Person#equals(Object) so
- * as to ensure that the client with exactly the same fields will be removed.
- *
- * Supports a minimal set of list operations.
- *
- * @see Person#isSamePerson(Person)
- */
-public class UniquePersonList implements Iterable<Person> {
-
-    private final ObservableList<Person> internalList = FXCollections.observableArrayList();
-    private final ObservableList<Person> internalUnmodifiableList =
-            FXCollections.unmodifiableObservableList(internalList);
-
-    /**
-     * Returns true if the list contains an equivalent client as the given argument.
-     */
-    public boolean contains(Person toCheck) {
-        requireNonNull(toCheck);
-        return internalList.stream().anyMatch(toCheck::isSamePerson);
-    }
-
-    /**
-     * Adds a client to the list.
-     * The client must not already exist in the list.
-     */
-    public void add(Person toAdd) {
-        requireNonNull(toAdd);
-        if (contains(toAdd)) {
-            throw new DuplicatePersonException();
-        }
-        internalList.add(toAdd);
-    }
-
-    /**
-     * Replaces the client {@code target} in the list with {@code editedPerson}.
-     * {@code target} must exist in the list.
-     * The client identity of {@code editedPerson} must not be the same as another existing client in the list.
-     */
-    public void setPerson(Person target, Person editedPerson) {
-        requireAllNonNull(target, editedPerson);
-
-        int index = internalList.indexOf(target);
-        if (index == -1) {
-            throw new PersonNotFoundException();
-        }
-
-        if (!target.isSamePerson(editedPerson) && contains(editedPerson)) {
-            throw new DuplicatePersonException();
-        }
-
-        internalList.set(index, editedPerson);
-    }
-
-    /**
-     * Removes the equivalent client from the list.
-     * The client must exist in the list.
-     */
-    public void remove(Person toRemove) {
-        requireNonNull(toRemove);
-        if (!internalList.remove(toRemove)) {
-            throw new PersonNotFoundException();
-        }
-    }
-
-    public void setPersons(UniquePersonList replacement) {
-        requireNonNull(replacement);
-        internalList.setAll(replacement.internalList);
-    }
-
-    /**
-     * Replaces the contents of this list with {@code persons}.
-     * {@code persons} must not contain duplicate persons.
-     */
-    public void setPersons(List<Person> persons) {
-        requireAllNonNull(persons);
-        if (!personsAreUnique(persons)) {
-            throw new DuplicatePersonException();
-        }
-
-        internalList.setAll(persons);
-    }
-
-    /**
-     * Returns the backing list as an unmodifiable {@code ObservableList}.
-     */
-    public ObservableList<Person> asUnmodifiableObservableList() {
-        return internalUnmodifiableList;
-    }
-
-    @Override
-    public Iterator<Person> iterator() {
-        return internalList.iterator();
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof UniquePersonList // instanceof handles nulls
-                        && internalList.equals(((UniquePersonList) other).internalList));
-    }
-
-    @Override
-    public int hashCode() {
-        return internalList.hashCode();
-    }
-
-    /**
-     * Returns true if {@code persons} contains only unique persons.
-     */
-    private boolean personsAreUnique(List<Person> persons) {
-        for (int i = 0; i < persons.size() - 1; i++) {
-            for (int j = i + 1; j < persons.size(); j++) {
-                if (persons.get(i).isSamePerson(persons.get(j))) {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-}
+//package seedu.address.model.client;
+//
+//import static java.util.Objects.requireNonNull;
+//import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
+//
+//import java.util.Iterator;
+//import java.util.List;
+//
+//import javafx.collections.FXCollections;
+//import javafx.collections.ObservableList;
+//import seedu.address.model.client.exceptions.DuplicatePersonException;
+//import seedu.address.model.client.exceptions.PersonNotFoundException;
+//
+///**
+// * A list of persons that enforces uniqueness between its elements and does not allow nulls.
+// * A client is considered unique by comparing using {@code Person#isSamePerson(Person)}. As such, adding and updating of
+// * persons uses Person#isSamePerson(Person) for equality so as to ensure that the client being added or updated is
+// * unique in terms of identity in the UniquePersonList. However, the removal of a client uses Person#equals(Object) so
+// * as to ensure that the client with exactly the same fields will be removed.
+// *
+// * Supports a minimal set of list operations.
+// *
+// * @see Person#isSamePerson(Person)
+// */
+//public class UniquePersonList implements Iterable<Person> {
+//
+//    private final ObservableList<Person> internalList = FXCollections.observableArrayList();
+//    private final ObservableList<Person> internalUnmodifiableList =
+//            FXCollections.unmodifiableObservableList(internalList);
+//
+//    /**
+//     * Returns true if the list contains an equivalent client as the given argument.
+//     */
+//    public boolean contains(Person toCheck) {
+//        requireNonNull(toCheck);
+//        return internalList.stream().anyMatch(toCheck::isSamePerson);
+//    }
+//
+//    /**
+//     * Adds a client to the list.
+//     * The client must not already exist in the list.
+//     */
+//    public void add(Person toAdd) {
+//        requireNonNull(toAdd);
+//        if (contains(toAdd)) {
+//            throw new DuplicatePersonException();
+//        }
+//        internalList.add(toAdd);
+//    }
+//
+//    /**
+//     * Replaces the client {@code target} in the list with {@code editedPerson}.
+//     * {@code target} must exist in the list.
+//     * The client identity of {@code editedPerson} must not be the same as another existing client in the list.
+//     */
+//    public void setPerson(Person target, Person editedPerson) {
+//        requireAllNonNull(target, editedPerson);
+//
+//        int index = internalList.indexOf(target);
+//        if (index == -1) {
+//            throw new PersonNotFoundException();
+//        }
+//
+//        if (!target.isSamePerson(editedPerson) && contains(editedPerson)) {
+//            throw new DuplicatePersonException();
+//        }
+//
+//        internalList.set(index, editedPerson);
+//    }
+//
+//    /**
+//     * Removes the equivalent client from the list.
+//     * The client must exist in the list.
+//     */
+//    public void remove(Person toRemove) {
+//        requireNonNull(toRemove);
+//        if (!internalList.remove(toRemove)) {
+//            throw new PersonNotFoundException();
+//        }
+//    }
+//
+//    public void setPersons(UniquePersonList replacement) {
+//        requireNonNull(replacement);
+//        internalList.setAll(replacement.internalList);
+//    }
+//
+//    /**
+//     * Replaces the contents of this list with {@code persons}.
+//     * {@code persons} must not contain duplicate persons.
+//     */
+//    public void setPersons(List<Person> persons) {
+//        requireAllNonNull(persons);
+//        if (!personsAreUnique(persons)) {
+//            throw new DuplicatePersonException();
+//        }
+//
+//        internalList.setAll(persons);
+//    }
+//
+//    /**
+//     * Returns the backing list as an unmodifiable {@code ObservableList}.
+//     */
+//    public ObservableList<Person> asUnmodifiableObservableList() {
+//        return internalUnmodifiableList;
+//    }
+//
+//    @Override
+//    public Iterator<Person> iterator() {
+//        return internalList.iterator();
+//    }
+//
+//    @Override
+//    public boolean equals(Object other) {
+//        return other == this // short circuit if same object
+//                || (other instanceof UniquePersonList // instanceof handles nulls
+//                        && internalList.equals(((UniquePersonList) other).internalList));
+//    }
+//
+//    @Override
+//    public int hashCode() {
+//        return internalList.hashCode();
+//    }
+//
+//    /**
+//     * Returns true if {@code persons} contains only unique persons.
+//     */
+//    private boolean personsAreUnique(List<Person> persons) {
+//        for (int i = 0; i < persons.size() - 1; i++) {
+//            for (int j = i + 1; j < persons.size(); j++) {
+//                if (persons.get(i).isSamePerson(persons.get(j))) {
+//                    return false;
+//                }
+//            }
+//        }
+//        return true;
+//    }
+//}
